Refactor page.tsx with using PortfolioList
